+++ magpie/ui/user/__init__.py
@@ -7,3 +7,4 @@ def includeme(config):
     from magpie.ui.user.views import UserViews
     LOGGER.info("Adding UI user...")
     config.add_route(UserViews.edit_current_user.__name__, "/ui/users/current")
+    config.scan()